refactor: migrate redaction digest
