--- serverless/redaction-digest/.eslintrc.js
@@ -1,23 +1,0 @@
-module.exports = {
-  root: false,
-  parser: '@typescript-eslint/parser',
-  plugins: ['@typescript-eslint'],
-  extends: [
-    'eslint:recommended', // Recommended ESLint rules
-    'plugin:@typescript-eslint/eslint-recommended', // Disables rules from `eslint:recommended` that are already covered by the TypeScript typechecker
-    'plugin:@typescript-eslint/recommended', // Recommended TypeScript rules
-    'prettier/@typescript-eslint', // Disables rules from `@typescript-eslint/recommended` that are covered by Prettier
-    'plugin:prettier/recommended', // Recommended Prettier rules
-  ],
-  parserOptions: {
-    sourceType: 'module',
-  },
-  env: {
-    node: true,
-  },
-  ignorePatterns: ['build', 'node_modules'],
-  rules: {
-    '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
-  },
-}
