--- serverless/redaction-digest/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "redaction-digest",
-  "version": "1.0.0",
-  "description": "Send redaction digest to users",
-  "main": "index.js",
-  "scripts": {
-    "build": "rm -rf build && tsc && cp -r src/assets build/assets",
-    "dev": "npm run build && tsc-watch --onSuccess \"npx lambda-local -l build/index.js -t 600 -h handler -e examples/event-cloudwatch-scheduled-event.js\"",
-    "test": "echo 'No tests yet for redaction-digest'"
-  },
-  "author": "",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/bluebird": "3.5.33",
-    "@types/convict": "5.2.2",
-    "@types/lodash": "4.14.165",
-    "@types/node": "14.14.7",
-    "@types/nodemailer": "6.4.0",
-    "@types/pg-connection-string": "2.0.0",
-    "@types/validator": "13.1.0",
-    "lambda-local": "1.7.4",
-    "tsc-watch": "4.2.9",
-    "typescript": "4.7.3"
-  },
-  "dependencies": {
-    "@sentry/node": "5.27.4",
-    "aws-sdk": "2.1359.0",
-    "axios": "0.27.2",
-    "convict": "6.2.4",
-    "dotenv": "8.2.0",
-    "lodash": "4.17.21",
-    "module-alias": "2.2.2",
-    "moment": "2.29.4",
-    "nodemailer": "6.6.2",
-    "pg": "8.5.1",
-    "pg-connection-string": "2.4.0",
-    "reflect-metadata": "0.1.13",
-    "sequelize": "6.29.1",
-    "sequelize-typescript": "2.1.3",
-    "source-map-support": "0.5.19"
-  },
-  "_moduleAliases": {
-    "@shared": "../../shared/build"
-  }
-}
