--- serverless/redaction-digest/src/utils/logger.ts
@@ -1,11 +1,0 @@
-export class Logger {
-  source: string
-
-  constructor(source: string) {
-    this.source = source
-  }
-
-  log(message?: string): void {
-    console.log(`${this.source}: ${message}`)
-  }
-}
