+++ sst/package.json
@@ -14,6 +14,7 @@
   },
   "devDependencies": {
     "@tsconfig/node16": "^16.1.0",
+    "@types/lodash": "^4.14.195",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "aws-cdk-lib": "2.84.0",
     "constructs": "10.1.156",
@@ -33,7 +34,14 @@
     "aws-cdk": "^2.87.0",
     "aws4-axios": "^3.2.24",
     "cronitor": "^2.3.3",
-    "dotenv": "^16.3.1"
+    "dotenv": "^16.3.1",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.4",
+    "pg": "^8.11.1",
+    "pg-connection-string": "^2.6.1",
+    "pg-hstore": "^2.3.4",
+    "sequelize": "^6.29.1",
+    "sequelize-typescript": "^2.1.3"
   },
   "lint-staged": {
     "*.{js,ts}": "eslint --cache --fix"
