+++ sst/packages/core/src/database/client.ts
@@ -1,5 +1,6 @@
 import { drizzle, PostgresJsDatabase } from 'drizzle-orm/postgres-js'
 import postgres from 'postgres'
+import { Sequelize } from 'sequelize-typescript'
 
 interface Options {
   max: number // max number of connections
@@ -14,4 +15,11 @@ export default class PostmanDbClient {
   public getClient() {
     return this.client
   }
+}
+
+export const getSequelize = (dbUri: string) => {
+  return new Sequelize(dbUri, {
+    dialect: 'postgres',
+    logging: false,
+  })
 }
