+++ sst/packages/functions/src/cron-jobs/api-key-expiry/main.ts
@@ -12,10 +12,9 @@ import { IS_LOCAL, LOCAL_DB_URI } from '../../env'
 import { reminderEmailMapper } from './helper'
 
 /*
-- This function is triggered by a cron job that runs every day at 12AM UTC, i.e. 8AM SGT
-- It picks up API keys that are expiring 28 days, 14 days, 3 days and 1 day from now
+- This selects API keys that are expiring 28 days, 14 days, 3 days and 1 day from now
 - DATE_TRUNC("day", valid_until) function round down the valid_until timestamp to the nearest day
-- The cron job will send emails to the API key owners to remind them to renew their API keys
+- An email is sent to email addresses associated with the API keys to remind them to renew their API keys
 */
 
 export async function handler() {
